tests: run tests in a single, per-process runtime
Seastar doesn't support running multiple instance of itself within a
single process, either in parallel or sequentially. Currently, tests
that use the `#[seastar::test]` annotation run their own seastar
instances but make sure to do sequentially - which kind of works because
an assertion fires after more than 256 shards are spawned within a
process.

This commit solves this problem by spawning a single seastar instance
for the whole process. When a test needs a seastar runtime, it will
lazily create it and cache access to it in a global variable. At the end
of the process (atexit), the runtime is stopped. This is similar to what
the original seastar framework does - the test harness creates a seastar
runtime at the start and shuts it down after everything ends.

The `#[seastar::test]` macro is adjusted to support the new scheme.